Move examples into single crate
