--- examples/animation/Cargo.toml
@@ -1,57 +1,0 @@
-[package]
-name = "rend3-animation-example"
-license = "MIT OR Apache-2.0 OR Zlib"
-version = "0.3.0"
-authors = ["The rend3 Developers"]
-edition = "2021"
-publish = false
-rust-version = "1.71"
-
-[lib]
-crate-type = ["lib", "cdylib"]
-
-[[bin]]
-name = "animation"
-path = "src/main.rs"
-
-[dependencies]
-# logging
-env_logger = { version = "0.10", default-features = false, features = ["auto-color", "humantime"] }
-# Linear algebra library
-glam = "0.24"
-# block on async functions
-pollster = "0.3"
-# Renderer core
-rend3 = { version = "^0.3.0", path = "../../rend3" }
-# Programmable render list that dictates how the scene renders
-rend3-routine = { version = "^0.3.0", path = "../../rend3-routine" }
-# Framework that deals with the event loop, setting up the renderer, and platform differences.
-rend3-framework = { version = "^0.3.0", path = "../../rend3-framework" }
-# Play animations on imported gltf models
-rend3-anim = { version = "^0.3.0", path = "../../rend3-anim" }
-# Import gltf models
-rend3-gltf = { version = "^0.3.0", path = "../../rend3-gltf" }
-# std::time::Instant that works on wasm
-web-time = "0.2"
-# windowing
-winit = "0.29.4"
-image = { version = "0.24", default-features = false, features = ["jpeg"] }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-console_log = "1"
-console_error_panic_hook = "0.1"
-js-sys = "0.3"
-web-sys = "0.3"
-wasm-bindgen = "0.2.83"
-wasm-bindgen-futures = "0.4"
-
-[target.'cfg(target_os = "android")'.dependencies]
-ndk-glue = { version = "0.7", features = ["logger"] }
-
-[package.metadata.android]
-build_targets = ["aarch64-linux-android"]
-
-[package.metadata.android.sdk]
-min_sdk_version = 28
-target_sdk_version = 28
-compile_sdk_version = 28
