--- examples/animation/src/main.rs
@@ -1,3 +1,0 @@
-fn main() {
-    rend3_animation_example::main();
-}
