--- examples/egui/Cargo.toml
@@ -1,42 +1,0 @@
-[package]
-name = "rend3-egui-example"
-license = "MIT OR Apache-2.0 OR Zlib"
-version = "0.3.0"
-authors = ["The rend3 Developers"]
-edition = "2021"
-publish = false
-rust-version = "1.71"
-
-[[bin]]
-name = "egui"
-path = "src/main.rs"
-
-[dependencies]
-# The egui immediate mode gui library
-egui = {git = "https://github.com/emilk/egui.git", branch="master"}
-# Winit integration with egui (turn off the clipboard feature)
-egui-winit = { git = "https://github.com/emilk/egui.git", branch="master", default-features = false, features = ["links", "wayland"] }
-# logging
-env_logger = { version = "0.10", default-features = false, features = ["auto-color", "humantime"] }
-# Linear algebra library
-glam = "0.24"
-# Importing png's
-image = { version = "0.24.3", default-features = false, features = ["png"] }
-# Renderer core
-rend3 = { version = "^0.3.0",path = "../../rend3" }
-# Egui renderer integration
-rend3-egui = { version = "^0.3.0",path = "../../rend3-egui" }
-# Framework that deals with the event loop, setting up the renderer, and platform differences.
-rend3-framework = { version = "^0.3.0", path = "../../rend3-framework" }
-# Programmable render list that dictates how the scene renders
-rend3-routine = { version = "^0.3.0",path = "../../rend3-routine" }
-# Windowing
-winit = "0.29.4"
-# Opening URL's
-webbrowser = "0.8.2"
-# Creating textures
-wgpu = "0.18"
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-# Winit integration with egui
-egui-winit = { git = "https://github.com/emilk/egui.git", branch="master", default-features = false, features = ["clipboard", "links", "wayland"] }
