--- examples/scene-viewer/Cargo.toml
@@ -1,85 +1,0 @@
-[package]
-name = "scene-viewer"
-version = "0.3.0"
-authors = ["The rend3 Developers"]
-edition = "2021"
-description = "A gltf scene viewer based on the rend3 3D rendering engine."
-readme = "README.md"
-repository = "https://github.com/BVE-Reborn/rend3"
-license = "MIT OR Apache-2.0 OR Zlib"
-keywords = ["3d", "graphics", "rend3", "renderer", "wgpu"]
-categories = [
-    "game-development",
-    "graphics",
-    "rendering",
-    "rendering::engine",
-    "wasm",
-]
-rust-version = "1.71"
-exclude = [
-    "bistro.jpg",
-    "emerald-square.jpg",
-    "scifi-base.jpg",
-    "screenshot.jpg",
-    "resources/default-scene",
-    "resources/rend3-default-scene.tar",
-]
-
-[lib]
-name = "rend3_scene_viewer_example"
-crate-type = ["lib", "cdylib"]
-
-
-[features]
-tracy = ["profiling/profile-with-tracy", "tracy-client"]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1"
-cfg-if = "1"
-glam = "0.24"
-# 0.7 is a completely unrelated library
-histogram = "0.6"
-image = { version = "0.24", default-features = false, features = [
-    "png",
-    "jpeg",
-    "tiff",
-] }
-indoc = "2"
-log = "0.4"
-pico-args = "0.5"
-pollster = "0.3"
-profiling = { version = "1", default-features = false }
-rend3 = { version = "^0.3.0", path = "../../rend3" }
-rend3-framework = { version = "^0.3.0", path = "../../rend3-framework" }
-rend3-gltf = { version = "^0.3.0", default-features = false, path = "../../rend3-gltf", features = [
-    "ddsfile",
-    "ktx2",
-] }
-rend3-routine = { version = "^0.3.0", path = "../../rend3-routine" }
-rustc-hash = "1"
-smallvec = "1"
-tracy-client = { version = "0.16", optional = true }
-web-time = "0.2"
-wgpu = "0.18.0"
-wgpu-profiler = "0.15.0"
-winit = "0.29.4"
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-wasm-bindgen-futures = "0.4"
-
-[target.'cfg(target_os = "android")'.dependencies]
-ndk-glue = { version = "0.7", features = ["logger"] }
-
-[package.metadata.android]
-build_targets = ["aarch64-linux-android"]
-assets = "resources"
-
-[package.metadata.android.sdk]
-min_sdk_version = 28
-target_sdk_version = 28
-compile_sdk_version = 28
-
-[package.metadata.android.application.activity]
-orientation = "landscape"
