--- examples/scene-viewer/src/main.rs
@@ -1,3 +1,0 @@
-fn main() {
-    rend3_scene_viewer_example::main()
-}
