--- examples/skinning/src/main.rs
@@ -1,3 +1,0 @@
-fn main() {
-    rend3_skinning_example::main();
-}
