+++ examples/src/animation/README.md
