+++ examples/src/animation/resources/LICENSE
