+++ examples/src/animation/resources/cube_3.gltf
