+++ examples/src/animation/resources/scene.bin
