+++ examples/src/animation/resources/scene.gltf
