+++ examples/src/animation/resources/textures/rabbit_baseColor.jpeg
