+++ examples/src/cube_no_framework/Cargo.toml
