+++ examples/src/egui/images/rust-logo-128x128-blk.png
