+++ examples/src/egui/screenshot.png
