+++ examples/src/main.rs
@@ -1,0 +1,3 @@
+fn main() {
+    rend3_examples::main();
+}
