+++ examples/src/scene_viewer/README.md
