+++ examples/src/scene_viewer/bistro.jpg
