+++ examples/src/scene_viewer/emerald-square.jpg
