+++ examples/src/scene_viewer/resources/skybox/bottom.jpg
