+++ examples/src/scene_viewer/resources/skybox/front.jpg
