+++ examples/src/scene_viewer/resources/skybox/left.jpg
