+++ examples/src/scene_viewer/resources/skybox/right.jpg
