+++ examples/src/scene_viewer/scifi-base.jpg
