+++ examples/src/scene_viewer/screenshot.jpg
