+++ examples/src/skinning/README.md
