+++ examples/src/skinning/RiggedSimple.glb
