+++ examples/src/static_gltf/README.md
