+++ examples/src/static_gltf/data.glb
