+++ examples/src/static_gltf/screenshot.png
