+++ examples/src/textured_quad/Cargo.toml
