+++ examples/src/textured_quad/README.md
