+++ examples/src/textured_quad/checker.png
