+++ examples/src/textured_quad/screenshot.png
