--- examples/static-gltf/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "rend3-static-gltf-example"
-license = "MIT OR Apache-2.0 OR Zlib"
-version = "0.3.0"
-authors = ["The rend3 Developers"]
-edition = "2021"
-publish = false
-rust-version = "1.71"
-
-[[bin]]
-name = "static-gltf"
-path = "src/main.rs"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-# logging
-env_logger = { version = "0.10", default-features = false, features = ["auto-color", "humantime"] }
-# Linear algebra library
-glam = "0.24"
-# parsing model files
-gltf = { version = "1.0", default-features = false, features = ["import", "utils"] }
-# Renderer core
-rend3 = { version = "^0.3.0", path = "../../rend3"}
-# Framework that deals with the event loop, setting up the renderer, and platform differences.
-rend3-framework = { version = "^0.3.0", path = "../../rend3-framework" }
-# Programmable render list that dictates how the scene renders
-rend3-routine = { version = "^0.3.0", path = "../../rend3-routine" }
-# windowing
-winit = "0.29.4"
