+++ rend3-framework/Cargo.toml
@@ -21,7 +21,7 @@ profiling = { version = "1", default-features = false }
 rend3 = { version = "0.3.0", path = "../rend3" }
 rend3-routine = { version = "0.3.0", path = "../rend3-routine" }
 thiserror = { version = "1" }
-winit = "0.29.4"
+winit = { version = "0.29.4", features = ["rwh_05"] }
 wgpu = "0.18.0"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
